Moved ClusteredSphereRenderer to mmstd_moldyn
