+++ plugins/mmstd_moldyn/src/mmstd_moldyn.cpp
@@ -44,6 +44,7 @@
 #include "rendering/ArrowRenderer.h"
 #include "rendering/GrimRenderer.h"
 #include "rendering/OracleSphereRenderer.h"
+#include "rendering/ClusteredSphereRenderer.h"
 
 #include "mmstd_moldyn/EllipsoidalDataCall.h"
 #include "mmstd_moldyn/ParticleGridDataCall.h"
@@ -152,6 +153,7 @@ namespace {
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::io::MMPGDWriter>();
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::io::MMPGDDataSource>();
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::rendering::OracleSphereRenderer>();
+            this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::rendering::ClusteredSphereRenderer>();
             // register calls here:
             this->call_descriptions.RegisterAutoDescription< ::megamol::stdplugin::moldyn::BrickStatsCall>();
             this->call_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::EllipsoidalParticleDataCall>();
